+++ packages/brick_offline_first/test/mixins/delete_all_mixin_test.dart
@@ -1,5 +1,5 @@
 import 'package:brick_offline_first/mixins.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'package:brick_sqlite/memory_cache_provider.dart';
 import 'package:brick_sqlite/sqlite.dart';
@@ -26,7 +26,6 @@ class DeleteAllRepository extends OfflineFirstWithTestRepository
 }
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
   sqfliteFfiInit();
 
   group('DeleteAllMixin', () {