+++ packages/brick_offline_first_with_graphql/lib/src/graphql_request_sqlite_cache_manager.dart
@@ -5,12 +5,12 @@ import 'dart:convert';
 import 'package:brick_offline_first/offline_queue.dart';
 import 'package:gql/language.dart' as lang;
 import 'package:gql_exec/gql_exec.dart';
-import 'package:sqflite/sqflite.dart';
+import 'package:sqflite_common/sqlite_api.dart' show DatabaseFactory;
 
 class GraphqlRequestSqliteCacheManager extends RequestSqliteCacheManager<Request> {
   GraphqlRequestSqliteCacheManager(
     String databaseName, {
-    DatabaseFactory? databaseFactory,
+    required DatabaseFactory databaseFactory,
     Duration? processingInterval,
     bool? serialProcessing,
   }) : super(