+++ packages/brick_offline_first_with_graphql/test/graphql_offline_request_queue_test.dart
@@ -1,13 +1,13 @@
 import 'package:brick_offline_first_with_graphql/src/graphql_offline_queue_link.dart';
 import 'package:brick_offline_first_with_graphql/src/graphql_offline_request_queue.dart';
 import 'package:brick_offline_first_with_graphql/src/graphql_request_sqlite_cache_manager.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:sqflite/sqlite_api.dart';
+import 'package:test/test.dart';
+import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 void main() {
   final offlineClient = GraphqlOfflineQueueLink(
-    GraphqlRequestSqliteCacheManager('db'),
+    GraphqlRequestSqliteCacheManager('db', databaseFactory: databaseFactoryFfi),
   );
 
   group('GraphqlOfflineRequestQueue', () {