+++ packages/brick_offline_first_with_graphql/test/graphql_offline_sqlite_cache_test.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:brick_offline_first/src/offline_queue/request_sqlite_cache.dart';
 import 'package:brick_offline_first_with_graphql/src/graphql_request_sqlite_cache.dart';
 import 'package:brick_offline_first_with_graphql/src/graphql_request_sqlite_cache_manager.dart';
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:gql_exec/gql_exec.dart';
 import 'package:gql/language.dart';
 import 'package:logging/logging.dart';
@@ -14,7 +14,6 @@ import 'package:sqflite_common/sqlite_api.dart';
 class MockLogger extends Mock implements Logger {}
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
   sqfliteFfiInit();
 
   group('GraphqlRequestSqliteCache', () {