+++ packages/brick_offline_first_with_rest/lib/offline_first_with_rest.dart
@@ -4,3 +4,4 @@ export 'package:brick_offline_first_with_rest_abstract/annotations.dart';
 export 'package:brick_rest/rest.dart' show FieldRename, Rest, RestProvider, RestSerializable;
 export 'package:brick_offline_first_with_rest/src/offline_first_with_rest_adapter.dart';
 export 'package:brick_offline_first_with_rest/src/offline_first_with_rest_repository.dart';
+export 'package:brick_offline_first_with_rest/src/offline_queue/rest_request_sqlite_cache_manager.dart';