+++ packages/brick_offline_first_with_rest/test/helpers/__mocks__.dart
@@ -1,6 +1,5 @@
 import 'package:brick_core/core.dart';
 import 'package:brick_offline_first_with_rest/offline_first_with_rest.dart';
-import 'package:brick_offline_first_with_rest/src/offline_queue/rest_request_sqlite_cache_manager.dart';
 import 'package:http/http.dart' as http;
 import 'package:brick_rest/rest.dart';
 import 'package:brick_sqlite/memory_cache_provider.dart';
@@ -51,7 +50,7 @@ class TestRepository extends OfflineFirstWithRestRepository {
           sqliteProvider: _sqliteProvider,
           memoryCacheProvider: MemoryCacheProvider([MemoryDemoModel]),
           migrations: {const DemoModelMigration()},
-          offlineQueueHttpClientRequestSqliteCacheManager: RestRequestSqliteCacheManager(
+          offlineQueueManager: RestRequestSqliteCacheManager(
             '$inMemoryDatabasePath/queue',
             databaseFactory: databaseFactoryFfi,
           ),