+++ packages/brick_offline_first_with_rest/test/offline_queue/rest_offline_requeust_queue_test.dart
@@ -1,12 +1,19 @@
 import 'package:brick_offline_first_with_rest/src/offline_queue/rest_offline_queue_client.dart';
 import 'package:brick_offline_first_with_rest/src/offline_queue/rest_offline_request_queue.dart';
 import 'package:brick_offline_first_with_rest/src/offline_queue/rest_request_sqlite_cache_manager.dart';
+import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'package:test/test.dart';
 
 import '__helpers__.dart';
 
 void main() {
-  final offlineClient = RestOfflineQueueClient(stubResult(), RestRequestSqliteCacheManager('db'));
+  final offlineClient = RestOfflineQueueClient(
+    stubResult(),
+    RestRequestSqliteCacheManager(
+      'db',
+      databaseFactory: databaseFactoryFfi,
+    ),
+  );
 
   group('RestOfflineRequestQueue', () {
     test('#start', () {