+++ packages/brick_sqlite/test/__mocks__/demo_model_adapter.dart
@@ -2,7 +2,7 @@ import 'package:brick_core/core.dart' show Query;
 import 'package:brick_sqlite/sqlite.dart';
 // ignore: unused_import, unused_shown_name
 import 'package:brick_sqlite_abstract/db.dart';
-import 'package:sqflite/sqflite.dart' show DatabaseExecutor;
+import 'package:sqflite_common/sqlite_api.dart' show DatabaseExecutor;
 
 import 'demo_model.dart';